pyproject.toml
[tool.poetry]
name = "catnotcatdetector"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
torch = "2.1.0"
torchvision = "0.16.0"
conda-lock = "^2.5.1"
dvc = "^3.36.0"
dvc-gdrive = "^3.0.0"
matplotlib = "^3.8.2"
scikit-learn = "^1.3.2"
pre-commit = "3.6.0"
pyarrow = "^14.0.2"
ipykernel = "^6.27.1"
wget = "^3.2"
pytest = "^7.4.4"
lightning = "^2.1.3"
hydra-core = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
